+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * enhancements
   * [CLI] Support `--app` option to start an application and its dependencies
+  * [Kernel] Introduce `Application.Behaviour` to define application module callbacks
 
 # v0.7.2 (2012-12-04)